requirements.txt
# Data processing
pandas>=2.0.0
numpy>=1.24.0
polars>=0.20.0

# API clients
requests>=2.31.0
httpx>=0.25.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2

# Deep Learning (optional)
torch>=2.0.0

# Uncertainty quantification
mapie>=0.7

# Visualization & Dashboard
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
